==> scripts/utilities/07-quality-check-scored-macro.R <==
## Horse-Cattle-Elk Grazing Interaction Study Rproj
## Step 4: Quality Check Scored Excel Scoring Macro (XLSM)

## What this script does:
## Reads in the completed excel macro (must be first saved as a csv)
## Randomly selects a percentage of the completed records for review
## Writes out a csv with these records to be copied into the excel macro
## this csv can then be copied and pasted into the excel macro (HorseImaging2018.xlsm) for scoring

## What this script requires:
## the csv file from the completed excel macro "BRT_06052019_10032019_subjects_chunk1.csv"
## this csv file needs to be located in a chunk folder within the collection folder
## this csv file is named by the photo collection folder and "~subjects_chunk[x]"
## e.g., BRT_06052019_10032019_subjects_chunk1.csv

# clear the R environment
rm(list=ls(all=TRUE))

# load in the required libraries
source("C:/Users/andre/Dropbox/Rproj/Horse-Cattle-Elk-Grazing-Interaction-Study/packages.R")
source("C:/Users/andre/Dropbox/Rproj/Horse-Cattle-Elk-Grazing-Interaction-Study/functions.R")

# print the current R version in the console to check if your R version matches mine (which is 4.0.3)
R.Version()

# print the session info to check which language locale is currently configured for this environment
# this is important because the locale sets the text file encoding on the OS
sessionInfo()

# C:\TEMP\quality-check\in-progress\A51_07122020_10112020\chunk2

# TODO setup the working directory using the environment.R script
# Define the location of the files
currentwd <- file.path("C:", "TEMP", "quality-check", "in-progress", "A51_07122020_10112020", "chunk2")

collection_folder <- "A51_07122020_10112020"

chunk_folder <- "chunk2"

# set the working directory to read in the files from the correct location on your hard drive (or on an external hard drive)
setwd(currentwd)

# double check the working directory to make sure its correct
getwd()

# read in the csv file that contains the metadata for all photos in the collection folder (e.g., BRL_06052019_07022019)
chunkcsv <- read.csv(paste0(paste(collection_folder, "subjects", chunk_folder, "completed", sep = "_"), ".csv"))

# if there are less than 50 rows in a chunk, than select at least 10 rows to check
# if there are more than 100 rows in a chunk, than select 30 rows to check
if(nrow(chunkcsv) < 100){
  sample_size <- 10
} else if (nrow(chunkcsv) >= 100 && (nrow(chunkcsv) < 300)){
  sample_size <- (nrow(chunkcsv) / 10)
} else{
  sample_size <- 30
  }

# select a random sample of the completed scoring macro
student_data <- slice_sample(chunkcsv, n = sample_size)

# arrange the student data in ascending order by the RecordNumber
# this will make it easier to score the photos if they are in order
student_data <- arrange(student_data, RecordNumber)

# print the record numbers to check which rows were selected
student_data$RecordNumber

# create a file name for the student data
excelfilename1 <- paste0(paste(collection_folder, "subjects", chunk_folder, "student_data", sep = "_"), ".csv")

# write out this file to the working directory
# this will overwrite any files with the same name
readr::write_csv(student_data, excelfilename1)

# remove the columns from the test data that contain student-generated data
# this data will be copied and pasted into a blank scoring macro
test_sample <- student_data[1:8]

# print the data frame to check that the rows selected match the rows from the student data
test_sample$RecordNumber

# create a file name for the test data
excelfilename2 <- paste0(paste(collection_folder, "subjects", chunk_folder, "test_sample", sep = "_"), ".csv")

# write out this file to the working directory
# this will overwrite any files with the same name
readr::write_csv(test_sample, excelfilename2)

# read in the completed excel macro containing data generated by you
expert_data <- read.csv(paste0(paste(collection_folder, "subjects", chunk_folder, "expert_data", sep = "_"), ".csv"))

# print the record numbers to make sure they match
expert_data$RecordNumber
test_sample$RecordNumber

# use a custom function to generate count columns for the data generated by you
expert_species <- allspecies(expert_data)

# use a custom function to generate count columns for the data generated by students
student_species <- allspecies(student_data)

# use a custom function to generate counts for each species for the data generated by you
expert_counts <- counts.df(expert_species)

# print this value to check its contents
expert_counts

# use a custom function to generate counts for each species for the data generated by students
student_counts <- counts.df(student_species)

# print this value to check its contents
student_counts

# TODO extract test photos to set/adjust date and time

# TODO drop photos of U of A personnel checking cameras, photos of ropes or climbers

# TODO copy photos with some ambiguity and needing a second opinion into a separate folder for manual review

# change the drive letter to G

# create new directory

# copy files to new directory

# copy data into blank macro
